+++ agent/src/test/java/com/teamscale/jacoco/agent/options/AgentOptionsTest.java
@@ -137,8 +137,8 @@ public class AgentOptionsTest {
 	/** Tests that supplying both revision and commit info is forbidden. */
 	@Test
 	public void testBothRevisionAndCommitSupplied() throws URISyntaxException {
-		String message = "'teamscale-revision' is incompatible with 'teamscale-commit', "
-				+ "'teamscale-commit-manifest-jar', or 'teamscale-git-properties-jar'.";
+		String message = "'teamscale-revision' is incompatible with 'teamscale-commit' and "
+				+ "'teamscale-commit-manifest-jar'.";
 
 		File jar = new File(getClass().getResource("manifest-and-git-properties.jar").toURI());
 
@@ -149,10 +149,6 @@ public class AgentOptionsTest {
 		assertThatThrownBy(
 				() -> getAgentOptionsParserWithDummyLogger().parse(
 						"teamscale-revision=1234,teamscale-commit-manifest-jar=" + jar.getAbsolutePath()))
-				.isInstanceOf(AgentOptionParseException.class).hasMessageContaining(message);
-		assertThatThrownBy(
-				() -> getAgentOptionsParserWithDummyLogger().parse(
-						"teamscale-revision=1234,teamscale-git-properties-jar=" + jar.getAbsolutePath()))
 				.isInstanceOf(AgentOptionParseException.class).hasMessageContaining(message);
 	}